Add the option to tag a project to a post.
This commit adds a command-line flag `projects_dir` which should point
to a folder containing project logs. When you select a project log from
the dropdown list on the edit page, the post is added to that project
log in addition to the full journal.

The full journal will show a `@project` tag at the start of the post.